requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
django-filter
# djangorestframework_simplejwt